feat(detector-aws): add ECS detector implementing DetectorSync interface
